security: define runAsUser, runAsGroup, and fsGroup
Potential attackers may gain access to a container and leverage its existing privileges to conduct an attack. Therefore, it is not recommended to deploy containers with root privileges unless it is absolutely necessary. This control identifies all the Pods running as root or can escalate to root. https://hub.armo.cloud/docs/c-0013