PLAT-636: create cronjob generation (#13)
* Add the cronjob definition

This file is not used by anything except of documentation for building the cronjob service

* Add the cronjob class

By extending the cdk8s API we can create our own classes as the CronJob class does not exist yet. Use the cronjob-definition.ts as documentation for this class

* Add the cronjob service

This will introduce the cronjob service type. Some refactor had to be done to support both the webservice and the cronjob by sharing code.

* Add testing of the cronjob service

* Remove an unfortunate import

* Add the schedule to the action inputs

* Update the build

* Update the fixture cluster name

* Add cronjob to the README

* Add nodegroup input

This field is a string that accepts the node group selector. Will remove the inputs.fargate when we have migrated the previews to inputs.nodegroup. inputs.instance is a fallback if the nodegroup is not defined. Also for previews

* Update dist folder

* Fix the cronjob nodegroup bug

* Update the dist folder